Use Redland RDF w/ Python bindings for RDF parsing
Doing a micro-benchmark, Redland parsing is about 16x faster compared to
pure-python rdflib. With rdflib, parsing used about 80% of the total
time during pushes to the API, with Redland it's only about 26% - about
as much as the database transactions.

Still have to figure out whether/how to deal with RDF syntaxes other
than N-Triples, but looks like a good choice for the performance at the
moment.